Move under tedge_api::measurement all related structs
