+++ crates/core/tedge_api/src/lib.rs
@@ -1,20 +1,14 @@
 pub mod alarm;
-pub mod builder;
-pub mod data;
 pub mod entity_store;
 pub mod error;
 pub mod event;
-pub mod group;
 pub mod health;
 pub mod measurement;
 pub mod messages;
 pub mod mqtt_topics;
-pub mod parser;
 pub mod path;
-pub mod serialize;
 mod software;
 mod store;
-pub mod utils;
 pub mod workflow;
 
 pub use download::*;
