+++ crates/core/tedge_api/src/measurement/data.rs
