+++ crates/core/tedge_api/src/measurement/serialize.rs
