+++ crates/core/tedge_api/src/measurement/utils.rs
