+++ crates/extensions/c8y_mapper_ext/src/json.rs
@@ -19,7 +19,7 @@ use crate::serializer;
 use clock::Clock;
 use clock::WallClock;
 use tedge_api::entity_store::EntityMetadata;
-use tedge_api::parser::*;
+use tedge_api::measurement::*;
 use time::OffsetDateTime;
 use time::{self};
